Move browser address implementation for Chrome into a platform-specific module.
